test: split formatter CLI tests
